Drops some metrics that can be derived from others
This is an attempt to optimise the metrics exposed to prometheus. In busy Jenkins instances, the number of time series exported can easily grow to significant sizes which can eventually lead to scraping timeouts. Eg: Multibranch organisations that will expose extra time series for each short-living branch.

The idea is to drop time series that can easily be replaced by other PromQL expression:
- `default_jenkins_builds_(success|failed)_build_count_*` can be replaced by using `default_jenkins_builds_duration_milliseconds_summary_count` and `status` label
- `default_jenkins_builds_last_build_result` can be replaced by `default_jenkins_builds_last_build_result_ordinal`